Recommendation System/travel_recommender/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
sqlite3
python-dotenv==1.0.0
faker==20.1.0
pydantic==2.5.0
scipy==1.11.4